+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Added "GetFileInfo()" to retrieve information on one or more file/directory paths.
   Variable length arguments and supports polymorphism with lists and dictionaries.
 - Added "Head()" and "Tail()" to return a subset of lists.
+- Added "RandomSample()" which returns a subset using either a percentage
+  or count of items.
 
 ### Changed